requirements.txt
pandas>=1.3.0
numpy>=1.21.0
matplotlib>=3.5.0
seaborn>=0.11.0
scikit-learn>=1.0.0
jupyter>=1.0.0
notebook>=6.4.0
ipywidgets>=7.6.0
tqdm>=4.62.0
requests>=2.25.0
beautifulsoup4>=4.9.0
nltk>=3.6.0
spacy>=3.4.0
transformers>=4.15.0
torch>=1.10.0
faiss-cpu>=1.7.0
chromadb>=0.3.0
openai>=0.27.0
langchain>=0.0.150
streamlit>=1.20.0
plotly>=5.0.0
wordcloud>=1.8.0
textstat>=0.7.0 